#880: refresh bricks in deployment banner
