+++ src/options/pages/installed/InstalledPage.scss
